refactor: 스타일 유틸리티 함수(getSizeStyles, getFontSize, getBorderRadius)들을 별도 파일로 분리 #49
